+++ src/Audio/TrackManager.cpp
@@ -158,7 +158,7 @@ Track * TrackManager::PlayTrackFromCategory(String category, bool loop){
 	List<String> songs;
 	Track * t = NULL;
 	int tries = 100;
-	srand(Timer::GetCurrentTimeMs());
+	srand((unsigned int)Timer::GetCurrentTimeMs());
 	while (t == NULL){
 		--tries;
 		int randomSong = rand() % tc->tracks.Size();
